Adding histogram support to parallel coords plot (port from fastbit branch).
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/branches/1.12RC/src@6386 18c085ea-50e0-402c-830e-de6fd14e8384

+++ plots/ParallelCoordinates/ParallelCoordinatesViewerPluginInfo.C
@@ -282,7 +282,7 @@ ParallelCoordinatesViewerPluginInfo::InitializePlotAtts(
         vector<ArgExpr*> *args = argsExpr ? argsExpr->GetArgs() : NULL;
         if (!args)
             return;
-        for (int i=0; i<args->size(); i++)
+        for (size_t i=0; i<args->size(); i++)
         {
             ExprNode *arg = (ExprNode*)((*args)[i]->GetExpr());
             if (arg->GetTypeName() == "List")
